move transifex out of app_manager
