+++ corehq/apps/app_manager/tests/test_generators.py
@@ -1,14 +1,14 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 from __future__ import unicode_literals
-import os
-import polib
 
+import os
 from io import open
 
+import polib
 from django.test.testcases import SimpleTestCase
 
-from corehq.apps.app_manager.app_translations.generators import Translation, PoFileGenerator
+from corehq.apps.translations.generators import Translation, PoFileGenerator
 
 translations = [
     Translation('hello', 'नमस्ते', [('occurrence-hello', '')], '0:occurrence-hello'),
