+++ corehq/apps/translations/const.py
@@ -1,2 +1,2 @@
-from __future__ import unicode_literals
+from __future__ import absolute_import
 MODULES_AND_FORMS_SHEET_NAME = "Modules_and_forms"
