+++ corehq/tabs/tabclasses.py
@@ -1,13 +1,17 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
-from six.moves.urllib.parse import urlencode
 
-from django.urls import reverse
 from django.conf import settings
 from django.http import Http404
+from django.urls import reverse
 from django.utils.html import escape, strip_tags
 from django.utils.safestring import mark_safe
 from django.utils.translation import ugettext_noop, ugettext as _, ugettext_lazy
+from django_prbac.utils import has_privilege
+from memoized import memoized
+from six.moves import map
+from six.moves.urllib.parse import urlencode
+
 from corehq import privileges, toggles
 from corehq.apps.accounting.dispatcher import AccountingAdminInterfaceDispatcher
 from corehq.apps.accounting.models import Invoice, Subscription
@@ -34,6 +38,7 @@ from corehq.apps.reports.dispatcher import ProjectReportDispatcher, \
     CustomProjectReportDispatcher
 from corehq.apps.reports.models import ReportConfig, ReportsSidebarOrdering
 from corehq.apps.smsbillables.dispatcher import SMSAdminInterfaceDispatcher
+from corehq.apps.translations.integrations.transifex.utils import transifex_details_available_for_domain
 from corehq.apps.userreports.util import has_report_builder_access
 from corehq.apps.users.models import AnonymousCouchUser
 from corehq.apps.users.permissions import (
@@ -53,17 +58,12 @@ from corehq.messaging.scheduling.views import (
 )
 from corehq.messaging.util import show_messaging_dashboard
 from corehq.motech.dhis2.view import Dhis2ConnectionView, DataSetMapView
-from corehq.motech.views import MotechLogListView
 from corehq.motech.openmrs.views import OpenmrsImporterView
+from corehq.motech.views import MotechLogListView
 from corehq.privileges import DAILY_SAVED_EXPORT, EXCEL_DASHBOARD
 from corehq.tabs.uitab import UITab
 from corehq.tabs.utils import dropdown_dict, sidebar_to_dropdown, regroup_sidebar_items
 from corehq.toggles import PUBLISH_CUSTOM_REPORTS
-from memoized import memoized
-from django_prbac.utils import has_privilege
-from six.moves import map
-
-from custom.icds.translations.integrations.utils import transifex_details_available_for_domain
 
 
 class ProjectReportsTab(UITab):
@@ -131,7 +131,6 @@ class ProjectReportsTab(UITab):
         """
         Return the url for the start of the report builder, or the paywall.
         """
-        from corehq.apps.hqwebapp.templatetags.hq_shared_tags import toggle_enabled
         from corehq.apps.userreports.views import ReportBuilderDataSourceSelect
         return reverse(ReportBuilderDataSourceSelect.urlname, args=[self.domain])
